# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  Task:
    properties:
      Complete:
        type: boolean
      CreatedAt:
        format: date-time
        type: string
      DeletedAt:
        format: date-time
        type: string
      Description:
        type: string
      ID:
        format: uint64
        type: integer
      PersonID:
        format: int64
        type: integer
      ProjectID:
        format: int64
        type: integer
      Title:
        type: string
      UpdatedAt:
        format: date-time
        type: string
    type: object
    x-go-package: github.com/psebaraj/gogetitdone/models
info:
  description: Documentation of the GoGetItDone endpoints of the backend
  title: GoGetItDone
  version: 1.0.0
paths:
...
  /changedescrition/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskDescription
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /changeproject/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskProject
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /changetitle/task/{id}:
    patch:
      consumes:
      - application/json
      operationId: changeTaskTitle
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /complete/task/{id}:
    patch:
      operationId: toggleTaskComplete
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "400":
          $ref: '#/responses/nil'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /create/task:
    post:
      consumes:
      - application/json
      operationId: createTask
      produces:
      - application/json
      responses:
        "201":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "507":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /delete/task/{id}:
    delete:
      operationId: deleteTask
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "404":
          $ref: '#/responses/nil'
        "500":
          $ref: '#/responses/nil'
      tags:
      - Task
...
  /task/{id}:
    get:
      operationId: getTask
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            $ref: '#/definitions/Task'
        "404":
          $ref: '#/responses/nil'
      tags:
      - Task
  /tasks:
    get:
      operationId: getTasks
      produces:
      - application/json
      responses:
        "200":
          description: Task
          schema:
            items:
              $ref: '#/definitions/Task'
            type: array
        "204":
          $ref: '#/responses/nil'
      tags:
      - Task
produces:
- application/json
...
schemes:
- http
swagger: "2.0"